requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.85.2
greenlet==2.0.0
h11==0.14.0
idna==3.4
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
